Remove refator code to reduce change scope
